adding lots of new features.

+++ assets/js/adminControllers.js
@@ -4,6 +4,64 @@
 // ADMIN CONTROLS
 // **********
 
+function KidFeedCtrl($scope, $rootScope, $cookieStore, $http) {
+  if(!$cookieStore.get('currentKid')) {
+    return $rootScope.navigate('fade','/');
+  } else if(!$cookieStore.get('currentKid').admin){
+    return $rootScope.navigate('fade','/chores');
+  }
+
+  $rootScope.modal = false;
+  $rootScope.pageName = 'kid-feed';
+  $rootScope.pageTitle = 'Children\'s Progress';
+  $rootScope.headerColor = 'red';
+  $rootScope.pageIcon = 'profile';
+
+  $http.get('/api/kids/')
+    .success(function(data, status, headers, config) {
+      $scope.kids = data;
+    });
+  // moment.calendar = {
+  //     lastDay : '[Yesterday]',
+  //     sameDay : '[Today]',
+  //     nextDay : '[Tomorrow]',
+  //     lastWeek : 'MMM D YYYY',
+  //     nextWeek : 'MMM D YYYY',
+  //     sameElse : 'MMM D YYYY'
+  // };
+
+  // var currentDate = moment(),
+  //     id = ($rootScope.currentKid) ? $rootScope.currentKid._id : null,
+  //     date = currentDate.unix(),
+  //     url = "/api/tasks/all/";
+  
+  // $scope.filterExpr = {done: false};
+  // console.log($scope.filterExpr);
+  // $scope.day = currentDate.format('dddd'),
+
+  // $scope.prevDay = function() {
+  //   currentDate.subtract('days',1);
+  //   $scope.day = currentDate.format('dddd'),
+  //   date = currentDate.unix();
+
+  //   getList($http, url, date);
+
+  // }
+  // $scope.nextDay = function() {
+  //   currentDate.add('days',1);
+  //   $scope.day = currentDate.format('dddd'),
+  //   date = currentDate.unix();
+
+  //   getList($http, url, date);
+  // }
+  
+  // function getList(agent, url, date) {
+  //   agent.get(url+date)
+  //     .success(function(data) {
+  //       console.log(data);
+  //     });
+  // }
+}
 function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
   if(!$cookieStore.get('currentKid')) {
     return $rootScope.navigate('fade','/');
@@ -27,11 +85,11 @@ function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
   console.log($scope.filterExpr);
   console.log($scope.feed);
   console.log($scope.admin);
-  $scope.day = currentDate.format('dddd');
+  $scope.day = currentDate.calendar();
 
   $scope.prevDay = function() {
     currentDate.subtract('days',1);
-    $scope.day = currentDate.format('dddd'),
+    $scope.day = currentDate.calendar(),
     date = currentDate.unix();
 
     getList($http, url, date);
@@ -39,7 +97,7 @@ function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
   }
   $scope.nextDay = function() {
     currentDate.add('days',1);
-    $scope.day = currentDate.format('dddd'),
+    $scope.day = currentDate.calendar(),
     date = currentDate.unix();
 
     getList($http, url, date);
@@ -47,12 +105,14 @@ function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
   $scope.approveTransaction = function(task, transactionId) {
     var URL = '/api/transaction/'+transactionId,
         approved = (!task.approved);
-        
-    if(task.done){
-      task.approved = (!task.approved);
+        task.approved = (!task.approved);
+    if(task.transactionId){
       $http.put(URL,{transaction: {approved: approved}})
         .success(function(data, status, headers, config) {
         });
+    } else {
+      task.approving = true;
+      $scope.transact(task);
     }
   }
   $scope.transact = function(task) {
@@ -61,7 +121,7 @@ function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
 
         console.log(task);
         
-    if(!task.approved){
+    if(!task.approved || task.approving){
       task.done = (!task.done);
       if(task.transactionId){
         $http.delete('/api/transaction/'+task.transactionId,{done: task.done}).
@@ -71,7 +131,7 @@ function ChoreFeedCtrl($scope, $rootScope, $cookieStore, $http) {
             totalCounts($scope.tasks)
           });
       } else {
-        $http.post('/api/transaction/',{date: date, _kid: task._kid, _task: task._id}).
+        $http.post('/api/transaction/',{date: date, _kid: task._kid, _task: task._id, approved: task.approved}).
           success(function(data, status, headers, config) {
             console.log(data);
             task.transactionId = data._id;
@@ -151,9 +211,8 @@ function AddChoreCtrl($scope, $rootScope, $cookieStore, $http) {
   }
 
   $scope.saveTask = function(task) {
-    console.log(task);
-    $http.post('/api/task', task).
-      success(function(data, status, headers, config) {
+    $http.post('/api/task', task)
+      .success(function(data, status, headers, config) {
         $scope.task = {};
         $rootScope.navigate('LR','/admin-chores')
       });
@@ -223,6 +282,21 @@ function AddRewardCtrl($scope, $rootScope, $cookieStore, $http) {
   $rootScope.pageTitle = 'Add Reward';
   $rootScope.headerColor = 'red';
   $rootScope.pageIcon = 'settings';
+
+  $scope.addReward = function(reward) {
+    if(reward.name && reward.value){
+      // reward.avatar = {color: 'blue', icon: 'ice-cream'};
+
+      $http.post('/api/reward', reward)
+        .success(function(data) {
+          console.log(data);
+        });
+    }
+    console.log(reward)
+
+  }
+
+
 }
 
 function EditChoreCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
@@ -245,6 +319,17 @@ function EditChoreCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
   $http.get('/api/task/'+taskId)
     .success(function(data, status, headers, config) {
       $scope.task = data;
+
+      for (var x in data.repeated){
+        console.log(x)
+        if(data.repeated[x]){
+          document.querySelector('span#'+x).classList.add('selected');
+        }
+      }
+      // for (var x in data.repeated){
+      //   console.log(x)
+      //   document.querySelector('span#'+x).classList.add('selected');
+      // }
       console.log($scope.task);
     });
 
@@ -253,6 +338,12 @@ function EditChoreCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
       $scope.kids = data;
     });
 
+  $scope.toggleDay = function(day) {
+
+    var toggled = document.getElementById(day).classList.toggle('selected');
+    $scope.task.repeated[day] = toggled; 
+
+  }
   $scope.saveTask = function(task) {
     console.log(task);
     $http.put('/api/task/'+taskId,{task: task})
@@ -261,11 +352,14 @@ function EditChoreCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
       });
   }
   $scope.deleteTask = function(taskId) {
-    console.log(taskId);
-    $http.delete('/api/task/'+taskId)
-      .success(function(data, status, headers, config) {
-        $rootScope.navigate('LR','/admin-chores');
-      });
+    var r = confirm("Are you sure you want to delete "+ $scope.task.name +"?");
+    if(r){
+      $http.delete('/api/task/'+taskId)
+        .success(function(data, status, headers, config) {
+          $rootScope.navigate('LR','/admin-chores');
+        });
+    }
+    
   }
 }
 function EditKidCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
@@ -299,17 +393,20 @@ function EditKidCtrl($scope, $rootScope, $cookieStore, $http, $routeParams) {
       });
   }
   $scope.deleteKid = function(kidId) {
-    $http.delete('/api/kid/'+kidId)
-      .success(function(data, status, headers, config) {
-        $rootScope.navigate('LR','/admin-kids');
-      });
+    var r = confirm("Are you sure you want to delete "+ $scope.kid.name +"?");
+    if(r){
+      $http.delete('/api/kid/'+kidId)
+        .success(function(data, status, headers, config) {
+          $rootScope.navigate('LR','/admin-kids');
+        });
+    }
   }
 }
 function EditRewardCtrl($scope, $rootScope, $cookieStore, $http) {
 }
 
 
-function AdminCtrl($scope, $rootScope, $cookieStore, $http, $location, $routeParams) {
+function AdminCtrl($scope, $rootScope, $cookieStore, $http, $location, $routeParams, storage) {
   if(!$cookieStore.get('currentKid')) {
     return $rootScope.navigate('fade','/');
   } else if(!$cookieStore.get('currentKid').admin){
@@ -323,7 +420,8 @@ function AdminCtrl($scope, $rootScope, $cookieStore, $http, $location, $routePar
   $rootScope.headerColor = 'red';
   $rootScope.pageIcon = 'settings';
 
-  $scope.rewardCount = 19;
+  $scope.user = storage.get('dd_user')
+  console.log($scope.user);
 
   $http.get('/api/kids')
     .success(function(data, status, headers, config) {
@@ -334,7 +432,38 @@ function AdminCtrl($scope, $rootScope, $cookieStore, $http, $location, $routePar
     .success(function(data, status, headers, config) {
         $scope.choreCount = data.length;
       });
+
+  $http.get('/api/rewards')
+    .success(function(data, status, headers, config) {
+        $scope.rewardCount = data.length;
+      });
   
+  $scope.editUsername = function() {
+    angular.element(document.getElementById('username')).toggleClass('hide');
+    angular.element(document.getElementById('usernameInput')).toggleClass('hide');
+    document.querySelector('#usernameInput input').focus();
+    document.querySelector('#usernameInput input').select();
+  }
+  $scope.saveUsername = function(username) {
+    $http.put('/api/user/username',{username: username})
+      .success(function(data, status, headers, config) {
+        var user = storage.get('dd_user');
+        user.name = data.username;
+        storage.set('dd_user',user);
+        angular.element(document.getElementById('username')).toggleClass('hide');
+        angular.element(document.getElementById('usernameInput')).toggleClass('hide');
+      })
+      .error(function(err) {
+        
+      });
+  }
+  $scope.editAvatar = function() {
+    
+  }
+  $scope.editPassword = function() {
+    
+  }
+
 }
 function AdminChoresCtrl($scope, $rootScope, $cookieStore, $http, $location, $routeParams) {
   if(!$cookieStore.get('currentKid')) {
@@ -397,9 +526,9 @@ function AdminRewardsCtrl($scope, $rootScope, $cookieStore, $http, $location, $r
   $rootScope.headerColor = 'red';
   $rootScope.pageIcon = null;
   
-  $http.get('/api/tasks/all')
+  $http.get('/api/rewards')
     .success(function(data, status, headers, config) {
-        $scope.tasks = data;
+        $scope.rewards = data;
       });
